rktio: patch rktio_process for "/bin/sh" on Guix
Racket provides the functions `system` and `process`,
which execute shell commands using `sh` (or `cmd` on Windows).
Racket assumes that `sh` can be found at "/bin/sh",
which is not necessarily true on Guix.

This patch adds a special case for "/bin/sh" to `rktio_process`,
the C function that implements the core of `system`, `process`,
and related Racket functions.

Guix should enable the special case by defining the C preprocessor
macro `GUIX_RKTIO_BIN_SH` with the path to `sh` in the store.
If:

    1. The `GUIX_RKTIO_BIN_SH` macro is defined; and

    2. `rktio_process` is called with the exact path "/bin/sh"; and

    3. The file specified by `GUIX_RKTIO_BIN_SH` exists;

then `rktio_process` will execute the file specified
by `GUIX_RKTIO_BIN_SH` instead of "/bin/sh".

Compared to previous attempts to patch the Racket sources,
making this change at the C level is both:

    - More comprehensive: it catches all attempts to execute "/bin/sh",
      without having to track down the source of every occurance; and

    - Less intrusive: by guarding the special case with a C preprocessor
      conditional and a runtime check that the file in the store exists,
      we make it much less likely that it will "leak" out of Guix.