net/server: close connection before deducting from max-concurrent
Previously, the connection supervisor thread would report to the
server thread that it was done *before* shutting down `client-cust`.
Add an extra custodian layer so that the supervisor can shut down
`client-cust` without terminating itself.

It might no longer be necessary to prevent `handle` from shutting
down `client-cust`, but keeping that layer seems easier to reason
about.

TODO: What should the parameters and custodian be for calls to
`make-timeout-evt`?