net/server: misc doc edits
Add a caveat for `accept` like the one for the `#:tcp@`
argument to `serve` from `web-server/web-server`.
See: https://github.com/racket/racket/pull/4621#discussion_r1154899135

Clarify that `make-timeout-evt` isn't responsible for actually shutting
down the custodian, just for telling the supervisor when to do so.
Also, move the docs for `max-concurrent` before those for
`make-timeout-evt` to match the order of arguments in the blue box.

Use `code:comment` where the examples use an insecure OpenSSL client
to make sure people who copy and paste the code don't overlook
the margin note.

Note that the ports example ignores the connection custodian.